backend/app/requirements.txt
fastapi
uvicorn
python-multipart
sentence-transformers
faiss-cpu
pypdf
